Version 0.2.0 - Post-crate generation updating of Cargo manifest, - Generation crate testing with self-dependency injection, - Generation utilities testing

+++ src/generate/crate_scaffolds.rs
@@ -1,6 +1,6 @@
 //! Set up target crate
 
-use crate::{cli::{Cli, Paths, SubCommands}};
+use crate::{cli::{Cli, Paths, SubCommands}, generate::{yamls}};
 use fs_err::{tokio as fs};
 use futures::{future::{TryFutureExt}};
 use strum::EnumProperty;
@@ -16,6 +16,7 @@ use std::{
 #[derive(Debug, Error)]
 pub enum CrateScaffoldingError{ 
   #[error(transparent)] IOError(#[from] IOError),
+  #[error(transparent)] YAMLGenerationError(#[from] yamls::YAMLGenerationError),
   #[error("Cannot scaffold in a directory that can't be confirmed as empty {0}: It's hella dangerous")] NonEmptyTargetDir(PathBuf),
   #[error("Could not find crate dir at {0}")] MissingCrateDir(PathBuf),
   #[error("Cargo init project at `{crate_dir}` failed with `{error_string}`  ")] CargoInitFailed{
@@ -28,7 +29,9 @@ async fn create_testing_folder(
   cli: &Cli,
 ) -> Result<(), CrateScaffoldingError> {
   let temp_dir_path = &cli.get_output_project_dir();
-  fs::remove_dir_all(&temp_dir_path).await?;
+  if temp_dir_path.is_dir() {
+    fs::remove_dir_all(&temp_dir_path).await?;
+  }
   fs::create_dir_all(&temp_dir_path).await?;
   Ok(())
 } 
@@ -50,6 +53,7 @@ async fn init_crate(
   cli: &Cli
 ) -> Result<(), CrateScaffoldingError> {
   let dir_path = &cli.get_output_project_dir();
+  // dbg!(cli);
   async { Ok(dir_path.is_dir()) }
     .and_then(|is_dir| async move {
       if !is_dir {
@@ -92,14 +96,19 @@ async fn init_crate(
 
 /// Do all crate scaffolding jobs
 pub async fn scaffold_crate(cli: &Cli) -> Result<(), CrateScaffoldingError> {
-  if let Some(SubCommands::TestGeneration { .. }) = cli.inner_cli.command.as_ref() {
+  let is_test = if let Some(SubCommands::TestGeneration { .. }) = cli.inner_cli.command.as_ref() {
     create_testing_folder(cli).await?;
+    true
   } else {
     create_crate_folder_and_check_empty(cli).await?;
-  }
+    false
+  };
   init_crate(cli).await?;
   setup_tree_in_crate(cli).await?;
   setup_git_in_crate(cli).await?;
+  if is_test {
+    yamls::create_testing_spec_file(cli).await?;
+  }
   Ok(())
 }
